feat!: repackaging project
BREAKING CHANGE: NuGet package is now published as `VisualOn.FeatureIDE` for `net641` and `net6.0`.
