+++ Directory.Build.props
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <Project>
     <PropertyGroup>
-        <MinVerMinimumMajorMinor>2.0</MinVerMinimumMajorMinor>
+        <MinVerMinimumMajorMinor>4.0</MinVerMinimumMajorMinor>
         <MinVerTagPrefix>v</MinVerTagPrefix>
         <SolutionDir>$(MSBuildThisFileDirectory)/</SolutionDir>
         <NoWarn>CA1014;CA1031;CA1065;CA1812;CA2225;SA0001;SA1404;SA1636;NU1701;$(NoWarn)</NoWarn>
@@ -16,10 +16,9 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <!-- <AdditionalFiles Include="$(MSBuildThisFileDirectory)/.stylecop.json" link="stylecop.json" /> -->
-        <!-- <PackageReference Include="MinVer" PrivateAssets="All" /> -->
+        <PackageReference Include="MinVer" PrivateAssets="All" />
+        <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All"/>
         <PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" PrivateAssets="All" />
-        <!--<PackageReference Include="StyleCop.Analyzers" PrivateAssets="All" />-->
     </ItemGroup>
 
     <Target Name="_AfterMinVer" AfterTargets="MinVer">
