+++ Directory.Packages.props
@@ -13,6 +13,8 @@
 
   <ItemGroup>
     <PackageVersion Include="IKVM" Version="8.4.5" />
+    <!--<PackageVersion Include="System.Runtime" Version="4.3.1" />-->
+    <PackageVersion Include="Validation" Version="2.5.51" />
   </ItemGroup>
 
   <ItemGroup>
