+++ codecov.yml
@@ -1,0 +1,6 @@
+comment: false
+codecov:
+  require_ci_to_pass: false
+  notify:
+    after_n_builds: 3
+    wait_for_ci: true
