+++ featureide.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.2.32602.215
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{23D1173E-D636-41E5-9B9D-2123BDD7A2FB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "de.ovgu.featureide.fm", "src\de.ovgu.featureide.fm\de.ovgu.featureide.fm.csproj", "{8FBEE9FC-42B7-48CE-B5CF-35D0C1D21EE0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VisualOn.FeatureIDE", "src\VisualOn.FeatureIDE\VisualOn.FeatureIDE.csproj", "{8FBEE9FC-42B7-48CE-B5CF-35D0C1D21EE0}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B95567D4-AA88-41AF-AA84-EC452AFE83AA}"
 	ProjectSection(SolutionItems) = preProject
@@ -15,7 +15,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{8F344BA8-EF69-42E0-84C6-9CA1128DAFE6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "de.ovgu.featureide.fm.test", "test\de.ovgu.featureide.fm.test\de.ovgu.featureide.fm.test.csproj", "{87431B37-B6C3-432D-800C-28D71357E020}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VisualOn.FeatureIDE.Test", "test\VisualOn.FeatureIDE.Test\VisualOn.FeatureIDE.Test.csproj", "{87431B37-B6C3-432D-800C-28D71357E020}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
