--- global.json
@@ -1,5 +1,0 @@
-{
-  "msbuild-sdks": {
-    "Microsoft.Build.NoTargets" : "3.7.0"
-  }
-}
