+++ src/VisualOn.FeatureIDE/VisualOn.FeatureIDE.csproj
@@ -1,0 +1,100 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net461;net6.0</TargetFrameworks>
+    <Product>FeatureIDE</Product>
+    <Copyright>VisualOn GmbH $([System.DateTime]::Now.ToString("yyyy"))</Copyright>
+    <Authors>VisualOn</Authors>
+    <Company>VisualOn GmbH</Company>
+    <PublishRepositoryUrl>true</PublishRepositoryUrl>
+    <IncludeBuildOutput>true</IncludeBuildOutput>
+    <Sat4JVersion>2.3.5</Sat4JVersion>
+    <!-- renovate: datasource=github-releases depName=featureide packageName=FeatureIDE/FeatureIDE -->
+    <FeatureIDEVersion>v3.9.1</FeatureIDEVersion>
+
+    <Title>FeatureIDE</Title>
+    <Description>
+      An extensible framework for feature-oriented software development
+
+      This is the IKVM version; which is to say that this is compiled from Java to .NET
+    </Description>
+    <PackageTags>IKVM.NET Java .NET SAT</PackageTags>
+    <PackageLicenseExpression>LGPL-3.0-only</PackageLicenseExpression>
+
+    <TargetsForTfmSpecificBuildOutput>$(TargetsForTfmSpecificBuildOutput);GetMyPackageFiles</TargetsForTfmSpecificBuildOutput>
+    <ResolveIkvmRuntimeAssemblyDependsOn>DownloadJars;$(ResolveIkvmRuntimeAssemblyDependsOn)</ResolveIkvmRuntimeAssemblyDependsOn>
+
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="IKVM" />
+    <PackageReference Include="Validation" />
+
+    <Using Include="Validation" />
+
+    <None Include="$(SolutionDir)\README.md" Pack="true" PackagePath="\"/>
+  </ItemGroup>
+
+  <Target Name="GetMyPackageFiles">
+    <ItemGroup>
+      <BuildOutputInPackage Include="$(OutputPath)org.sat4j.core.dll" />
+      <BuildOutputInPackage Include="$(OutputPath)org.sat4j.pb.dll" />
+      <BuildOutputInPackage Include="$(OutputPath)de.ovgu.featureide.fm.dll" />
+    </ItemGroup>
+  </Target>
+
+  <Target Name="DownloadJars" BeforeTargets="Build">
+    <DownloadFile
+      SourceUrl="https://repository.ow2.org/nexus/content/repositories/releases/org/ow2/sat4j/org.ow2.sat4j.core/$(Sat4JVersion)/org.ow2.sat4j.core-$(Sat4JVersion).jar"
+      DestinationFolder="$(MSBuildProjectDirectory)\bin"
+      DestinationFileName="org.sat4j.core.jar"
+      SkipUnchangedFiles="true">
+    </DownloadFile>
+    <DownloadFile
+      SourceUrl="https://repository.ow2.org/nexus/content/repositories/releases/org/ow2/sat4j/org.ow2.sat4j.pb/$(Sat4JVersion)/org.ow2.sat4j.pb-$(Sat4JVersion).jar"
+      DestinationFolder="$(MSBuildProjectDirectory)\bin"
+      DestinationFileName="org.sat4j.pb.jar"
+      SkipUnchangedFiles="true">
+    </DownloadFile>
+    <DownloadFile
+      SourceUrl="https://github.com/FeatureIDE/FeatureIDE/releases/download/$(FeatureIDEVersion)/de.ovgu.featureide.lib.fm-$(FeatureIDEVersion).jar"
+      DestinationFolder="$(MSBuildProjectDirectory)\bin"
+      DestinationFileName="de.ovgu.featureide.fm.jar"
+      SkipUnchangedFiles="true">
+    </DownloadFile>
+
+    <ItemGroup>
+      <IkvmReference Include="$(MSBuildProjectDirectory)\bin\org.sat4j.core.jar">
+        <AssemblyName>org.sat4j.core</AssemblyName>
+        <AssemblyVersion>$(Sat4JVersion.Split(`.`)[0]).0.0.0</AssemblyVersion>
+        <AssemblyFileVersion>$(Sat4JVersion).0</AssemblyFileVersion>
+        <KeyFile>$(SolutionDir)featureide.snk</KeyFile>
+        <Debug>false</Debug>
+        <DisableAutoAssemblyName>true</DisableAutoAssemblyName>
+        <DisableAutoAssemblyVersion>true</DisableAutoAssemblyVersion>
+      </IkvmReference>
+      <IkvmReference Include="$(MSBuildProjectDirectory)\bin\org.sat4j.pb.jar">
+        <AssemblyName>org.sat4j.pb</AssemblyName>
+        <AssemblyVersion>$(Sat4JVersion.Split(`.`)[0]).0.0.0</AssemblyVersion>
+        <AssemblyFileVersion>$(Sat4JVersion)</AssemblyFileVersion>
+        <KeyFile>$(SolutionDir)featureide.snk</KeyFile>
+        <References>$(MSBuildProjectDirectory)\bin\org.sat4j.core.jar</References>
+        <Debug>false</Debug>
+        <DisableAutoAssemblyName>true</DisableAutoAssemblyName>
+        <DisableAutoAssemblyVersion>true</DisableAutoAssemblyVersion>
+      </IkvmReference>
+      <IkvmReference Include="$(MSBuildProjectDirectory)\bin\de.ovgu.featureide.fm.jar">
+        <AssemblyName>de.ovgu.featureide.fm</AssemblyName>
+        <AssemblyVersion>$(FeatureIDEVersion.Replace(`v`, ``).Split(`.`)[0]).0.0.0</AssemblyVersion>
+        <AssemblyFileVersion>$(FeatureIDEVersion.Replace(`v`, ``))</AssemblyFileVersion>
+        <KeyFile>$(SolutionDir)featureide.snk</KeyFile>
+        <References>$(MSBuildProjectDirectory)\bin\org.sat4j.core.jar;$(MSBuildProjectDirectory)\bin\org.sat4j.pb.jar</References>
+        <Debug>false</Debug>
+        <DisableAutoAssemblyName>true</DisableAutoAssemblyName>
+        <DisableAutoAssemblyVersion>true</DisableAutoAssemblyVersion>
+      </IkvmReference>
+    </ItemGroup>
+  </Target>
+
+</Project>
