--- test/de.ovgu.featureide.fm.test/Usings.cs
@@ -1,5 +1,0 @@
-global using NExpect;
-global using NSubstitute;
-global using NSubstitute.Extensions;
-global using NUnit.Framework;
-global using static NExpect.Expectations;
